misc:stml4xx: Extend SensorHub WAKE_SENSORS array to 24 bits -- L4.
All 16 sensor wake flags have been used. We need to expand the number of
available bits to implement new gestures.

Reviewed-on: http://gerrit.mot.com/729511
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ include/linux/motosh.h
@@ -537,7 +537,7 @@ extern long motosh_time_delta;
 
 extern unsigned int motosh_irq_disable;
 
-extern unsigned short motosh_g_wake_sensor_state;
+extern unsigned long motosh_g_wake_sensor_state;
 
 extern unsigned char stat_string[];
